// part9/courses/my-app/src/types/types.ts
interface CoursePartBase {
    name: string;
    exerciseCount: number;
}

interface CoursePartWithDescription extends CoursePartBase {
    description: string;
}

interface CoursePartBasic extends CoursePartWithDescription {
    kind: "basic"
}

interface CoursePartGroup extends CoursePartBase {
    groupProjectCount: number;
    kind: "group"
}

interface CoursePartBackground extends CoursePartWithDescription {
    backroundMaterial: string;
    kind: "background"
}


export type CoursePart = CoursePartBasic | CoursePartGroup| CoursePartBackground